server/requirements.txt
fastapi==0.108.0
uvicorn==0.25.0
python-multipart==0.0.6
pillow==10.1.0
pydantic==2.5.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
requests==2.31.0
